Some optimization changes and code cleanup
Changed blocks the player should spawn on to a blacklist of banned blocks as opposed to a whitelist of accepted blocks.
Now using a HashSet of materials to check blocks instead of a list of strings.
Moved logic to find a safe spawn location to a separate method.
Updated Javadoc comments to be appropriate to changes.